requirements.txt
flask
python-dotenv
flask_sqlalchemy
flask_login
SQLAlchemy
Werkzeug
Pillow
requests
uuid